build: compile host tests with -Wall -Wextra -Wconversion
Closes the last Tier 0 item in docs/testing-remediation.md. The test targets
previously set no warning flags at all.

Placed after FetchContent_MakeAvailable(googletest) so vendored GoogleTest,
which does not build clean under -Wconversion, is unaffected. Warnings rather
than -Werror: CI surfaces them, and a hard failure on a toolchain bump is
worse than a visible warning. First-party warning count is now zero across
all three suites, which is the state that makes a new one worth reading.

Honest result: -Wconversion surfaced no live bugs. All 13 conversion sites
were provably in range - masked byte writes (x & 0xFF) in protocol.cpp, a
uint32_t diagnostic counter in frame_scanner, a run count bounded by its
array, and test-file literals. The 28 -Wunused-parameter hits were eleven
empty no-op message handlers, a log-only handler whose macro compiles away,
and a deliberately-unimplemented CommInterfaceImpl stub - not handlers
ignoring a length they should have checked, which was the shape worth
looking for.

Two findings did come out of it:

- file_browser_test's NavigateToPath computed a bool result and never
  asserted on it, so the function's contract in the mocked environment was
  untested and a change to it could not fail anything. Now pinned with
  EXPECT_FALSE, matching the comment that had claimed the behaviour without
  checking it.
- The flag cannot have caught the bug that motivated it. 1d16237's size_t ->
  uint16_t narrowing lives in inter_mcu.cpp, which the ESP32 test build
  explicitly excludes, so this protects shared and test-compiled code only.
  Extending warnings to the two firmware builds is the follow-on and is a
  larger job, since ESP-IDF and the STM32 HAL headers do not build clean
  under -Wconversion.

Mechanical fixes rather than casts at every call site where possible:
FlatParams now takes size_t so the ~25 vector::size() call sites narrow once
inside it, and a VoiceAt helper does the int -> uint8_t narrowing after the
caller has already asserted the index is non-negative. -Wno-unused-parameter
is scoped to esp32_mocks.cpp alone, where matching vendor signatures with
unused parameters is the entire point of a stub; -Wconversion still applies
there, because a narrowing bug in a mock produces a wrong test result.

+++ firmware/daisy/src/memory.h
@@ -415,7 +415,7 @@ class LargeExtentPool {
                 free_runs_[w++] = cur;
             }
         }
-        run_count_ = w;
+        run_count_ = static_cast<uint16_t>(w);  // bounded by free_runs_ capacity
     }
 
     uint32_t free_bytes() const {
